Redefined state and state trigger hash set rules.

+++ src/main/java/com/ani/anole/domain/state/StateGroup.java
@@ -6,11 +6,11 @@ import java.util.Set;
  * Created by zsl on 17-3-6.
  */
 public class StateGroup {
-    public Long groupId;
+    public Integer groupId;
     public String name;
     public Set<String> tags;
 
-    public StateGroup(Long groupId, String name, Set<String> tags) {
+    public StateGroup(Integer groupId, String name, Set<String> tags) {
         this.groupId = groupId;
         this.name = name;
         this.tags = tags;
